build: enable MSBuild's warnings as errors
It turns out that the TreatWarningsAsErrors:true setting we
already have for our builds does not apply to MSBuild warnings
(just F# warnings) because I kept seeing these warnings in our
build logs without the build being broken. More info here[1].

Then one day I stumbled upon a **crash** (!!) that could have
been prevented if we had these warnings be considered errors,
because they were signaling an FSharp.Core versioning conflict.
More info in the PR that was causing this crash: [2].

Now, please note, that while this commit is landing in the
`stable` branch first and not causing any issue here, the
interesting changes will come when we merge this in to the
`master` branch, where the Xamarin.Forms projects live, and
where we need tweaks to make the build work after this setting
is enabled.

[1] https://stackoverflow.com/a/67676530/544947
[2] https://github.com/nblockchain/geewallet/pull/262