Frontend.XF: fix NETSDK1023 (warning as) error
Build error was:
```
/usr/lib/mono/msbuild/Current/bin/Sdks/Microsoft.NET.Sdk/targets/Microsoft.NET.Sdk.DefaultItems.targets(257,5):
error NETSDK1023: A PackageReference for 'NETStandard.Library' was included in your project. This package is
implicitly referenced by the .NET SDK and you do not typically need to reference it from your project.
For more information, see https://aka.ms/sdkimplicitrefs
[/__w/geewallet/geewallet/src/GWallet.Frontend.XF/GWallet.Frontend.XF.fsproj]
```

This is related to the new MSBuild warnings enabled in:
25297f77dacfbb8c5d884956513fd7fc60fcb66b