Frontend.XF.GTK: ignore MSB3276 for now
We choose to ignore this one for now before fixing it because
it doesn't seem to cause a real issue at runtime for now.

This (warning as) error surfaced after this change[1] was merged
to master[2].

Error is:
```
Found conflicts between different versions of the same dependent assembly. Please set the "AutoGenerateBindingRedirects" property to true in the project file.
For more information, see http://go.microsoft.com/fwlink/?LinkId=294190
```

[1] 25297f77dacfbb8c5d884956513fd7fc60fcb66b
[2] 7ca27a7f144127f090c235239cc4b665e5107a48