scripts/make.fsx: gen normal nupkg & prerelease
This way we catch potential issues with the non-prerelease nuget
package before the git tag is pushed.

For example, the following warning:

```
/usr/lib/dotnet/sdk/6.0.127/Sdks/NuGet.Build.Tasks.Pack/build/NuGet.Build.Tasks.Pack.targets(221,5): error NU5104: A stable release of a package should not have a prerelease dependency. Either modify the version spec of
dependency "Fsdk [0.6.0--date20230530-1155.git-3bb8d08, )" or update the version field in the nuspec. [/__w/geewallet/geewallet/src/GWallet.Backend/GWallet.Backend.fsproj]
```

So let's enable 'warnaserror' in the 'dotnet pack' command
if the version to pack is even (e.g. 0.8) and ignore the
warning for now if it's odd (e.g. 0.7) like it currently
happens. Because even numbers map to "stableness" and odd
numbers to the opposite.