Backend/FiatValueEstimation: add BitPay API
To increase reliability & accuracy for the USD exchange rates,
because from my testing I think there's one of the previous
providers (CoinGecko or CoinCap) I could gather that one of
them is very fast to reply with an error message complaining
about how fast/often we are querying them.

This commit is similar to the one that recently landed in the
master branch [1] adding CoinDesk, but this one goes to the
stable branch first because it provides rates for many
currencies, not just BTC (from my testing it seems it provides
rates for all of the ones that geewallet supports, except ETC
and SAI).

[1] 3114fc3ccb1b9d13deed02de64d6fc59ddd544ea