Frontend.XF: finish switching to main Grid layout
Use Grid instead of StackLayout as root layout in all places
left in the UI (as a last step after other previous commits
such as [1] and [2]) so that in Maui the layout is the same
as in Xamarin.Forms (was not true with StackLayout, see [3]).

NOTE: used several workarounds in code to make GTK version
behave more or less as before (Xamarin.Forms Grid layout
behaves somewhat differently on Android and GTK).

[1] 55a3146592f78e79544ea1c9313cb19b2ca53a15
[2] be1a275754101c00bbcc4328488c65fb5ca63d28
[3] https://github.com/dotnet/maui/wiki/Xamarin.Forms----MAUI.Controls-Layout-Differences#stacklayout