Frontend.XF: use Grid layout as root layout
Use Grid instead of StackLayout as root layout so that in Maui
the layout is the same as in Xamarin.Forms (was not true with
StackLayout).

Used several workarounds in code to make GTK version behave more
or less as before (Xamarin.Forms Grid layout behaves somewhat
differently on Android and GTK).