Backend/FiatValueEst: fix avg calc's edge case
While working on the unit tests for PR 278 [1], I noticed
that when I implemented this function in the stable branch
to calculate the average between 3 values discarding the
outlier, I missed the edge case when the distance between
the highest and the lowest is the same.

So, in this case, let's discard the intermediate (the
intermediate will in the end be the average, but that
is calculated later in the code).

[1] https://github.com/nblockchain/geewallet/pull/278