Backend/FeeRateEstimation: refactor (def priority)
Fee rate estimation in geewallet has always aimed to be high
because we've always aimed to focus on finishing L2 support,
so for channels to be opened/closed we wanted fast conf times.
However, LN development is taking a bit long, and we're now
sometimes facing periods of high-fee environments, so it's
interesting to start working on this feature.

Right now, this refactoring introduces two kinds of priority
but the UX towards the user doesn't change because the wallet
is still configured to choose the highest. This just lays the
groundwork for a future feature.