scripts/configure.fsx: clarify Frontend
When XamarinForms frontend can be built, the Console frontend
is still built as well, so here we're changing the frontend
elemenet from the configuration summary to be a list instead
of a single value. (Also DUs instead of strings.)

Further, it's still important to highlight which frontend will
be the default one (e.g. that will get run by the 'make run'
target). So we add a new "Default Frontend" element in the
summary.

This will be especially useful for an upcoming change that
will add MAUI frontend on top of XamarinForms (so, both XF and
MAUI frontend are built, but only one is default).