Backend: don't check eth tx after sending
After sending Ethereum transaction, don't poll the server to
check if transaction is out of gas. Eth transactions can now
take longer time to be confirmed and that poll could time out.

Instead, check if recipient is a contract address. If it is,
raise an exception (as sending to contract addresses is not yet
supported).